Fix for DIRSTUDIO-917 (Exporting for ApacheDS an AT with a not-set syntax length results in 'm-length: 0' output in the resulting LDIF file).
git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@1486628 13f79535-47bb-0310-9956-ffa450edef68

+++ ldap/schema/converter/src/main/java/org/apache/directory/api/converter/schema/AttributeTypeHolder.java
@@ -62,7 +62,7 @@ public class AttributeTypeHolder extends SchemaElementImpl
     private String syntax;
 
     /** The optional length for this attribute */
-    private int oidLen = -1;
+    private long oidLen = -1;
 
     /** The attribute uase. Default to userApplication */
     private UsageEnum usage = UsageEnum.USER_APPLICATIONS;
@@ -282,7 +282,7 @@ public class AttributeTypeHolder extends SchemaElementImpl
      * 
      * @return The attributeType's syntax length
      */
-    public int getOidLen()
+    public long getOidLen()
     {
         return oidLen;
     }
@@ -293,7 +293,7 @@ public class AttributeTypeHolder extends SchemaElementImpl
      * 
      * @param oidLen The attributeType's syntax length
      */
-    public void setOidLen( int oidLen )
+    public void setOidLen( long oidLen )
     {
         this.oidLen = oidLen;
     }
